Refactor: Apply dependency injection
dependency injection about bottomsheet
